- Revert back to single neuron output - Use BinaryAccuracy - Change target type to short as it should be

+++ dfpl/settings.py
@@ -33,7 +33,7 @@ ac_fp_compressed_numpy_type = np.bool8
 
 nn_fp_numpy_type = np.float32
 nn_fp_compressed_numpy_type = np.float32
-nn_target_numpy_type = np.float32
+nn_target_numpy_type = np.short
 
 nn_multi_fp_numpy_type = np.float32
 nn_multi_fp_compressed_numpy_type = np.float32
